moved duplicate notification check function from auxillary.py to services.notifications
