+++ Webapp/sources/services/notifications.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import List, Optional
 from datetime import datetime
 
 from flask_login import current_user
@@ -110,3 +110,70 @@ def add_user_by_email_request(person: models.Person, workgroup: models.WorkGroup
             wb="",
             wg_request="added",
         )
+
+
+def duplicate_notification_check(
+    people_ls: List[models.Person],
+    request_type: str,
+    status,
+    WG: str,
+    WB: models.WorkBook = None,
+) -> bool:
+    """
+    Check if the request has already been made - if someone already has received your request they will be
+    removed from the list of people to be sent the request
+    """
+    refined_people_ls = []
+    for person in people_ls:
+        if WB:
+            check = (
+                db.session.query(models.Notification)
+                .filter(models.Notification.wb == WB and models.Notification.wg == WG)
+                .filter(models.Notification.status == status)
+                .join(models.WBStatusRequest)
+                .join(models.Person, models.WBStatusRequest.person == models.Person.id)
+                .filter(models.Person.id == person.id)
+                .first()
+            )
+        else:
+            if request_type == "New Workgroup Membership Request":
+                # check for person applying to join workgroup - 1 PI may have many requests from different users
+                check = (
+                    db.session.query(models.Notification)
+                    .filter(models.Notification.wg == WG)
+                    .join(models.WGStatusRequest)
+                    .join(
+                        models.Person, models.WGStatusRequest.person == models.Person.id
+                    )
+                    .filter(models.Person == person.id)
+                    .first()
+                )
+            elif request_type == "New Workgroup Role Reassignment Request":
+                check = (
+                    db.session.query(models.Notification)
+                    .filter(models.Notification.wg == WG)
+                    .filter(models.Notification.status == status)
+                    .join(models.WGStatusRequest)
+                    .join(
+                        models.Person, models.WGStatusRequest.person == models.Person.id
+                    )
+                    .filter(models.Person.id == person.id)
+                    .first()
+                )
+            else:
+                check = (
+                    db.session.query(models.Notification)
+                    .filter(models.Notification.wg == WG)
+                    .filter(models.Notification.status == status)
+                    .filter(models.Notification.type == request_type)
+                    .join(models.WGStatusRequest)
+                    .join(
+                        models.Person, models.WGStatusRequest.person == models.Person.id
+                    )
+                    .filter(models.Person.id == person.id)
+                    .first()
+                )
+
+        if check is None:
+            refined_people_ls.append(person)
+    return not refined_people_ls
